Replace casting numbers to unsigned with shorthand

+++ rttest/test/test_api.cpp
@@ -53,8 +53,8 @@ TEST(TestApi, read_args_get_params) {
   EXPECT_EQ(params.update_period.tv_nsec, 50000);
   EXPECT_EQ(params.sched_priority, 42);
   EXPECT_EQ(params.sched_policy, static_cast<uint>(SCHED_FIFO));
-  EXPECT_EQ(params.stack_size, static_cast<uint>(102400));
-  EXPECT_EQ(params.prefault_dynamic_size, static_cast<uint>(102400));
+  EXPECT_EQ(params.stack_size, 102400u);
+  EXPECT_EQ(params.prefault_dynamic_size, 102400u);
   EXPECT_EQ(strcmp(params.filename, "foo.txt"), 0);
   EXPECT_EQ(0, rttest_finish());
 }
@@ -95,7 +95,7 @@ TEST(TestApi, init) {
   struct rttest_params params;
   EXPECT_EQ(0, rttest_get_params(&params));
 
-  EXPECT_EQ(params.iterations, static_cast<uint>(4321));
+  EXPECT_EQ(params.iterations, 4321u);
   EXPECT_EQ(params.update_period.tv_sec, update_period.tv_sec);
   EXPECT_EQ(params.update_period.tv_nsec, update_period.tv_nsec);
   EXPECT_EQ(params.sched_priority, 42);
@@ -121,7 +121,7 @@ TEST(TestApi, spin_once) {
   EXPECT_EQ(0, rttest_spin_once(test_callback, static_cast<void *>(&counter), &start_time, 0));
   // Block for longer than the update period
   clock_nanosleep(CLOCK_MONOTONIC, 0, &update_period, NULL);
-  EXPECT_EQ(counter, static_cast<uint>(1));
+  EXPECT_EQ(counter, 1u);
 
   EXPECT_EQ(0, rttest_finish());
 }
